+++ src/frontend/points.c
@@ -3,12 +3,12 @@ Copyright 1990 Regents of the University of California.  All rights reserved.
 Author: 1985 Wayne A. Christopher, U. C. Berkeley CAD Group 
 **********/
 
-#include <ngspice.h>
-#include <cpdefs.h>
-#include <ftedefs.h>
-#include <dvec.h>
-#include <graph.h>
-#include <ftedbgra.h>
+#include <ngspice/ngspice.h>
+#include <ngspice/cpdefs.h>
+#include <ngspice/ftedefs.h>
+#include <ngspice/dvec.h>
+#include <ngspice/graph.h>
+#include <ngspice/ftedbgra.h>
 
 #include "points.h"
 
@@ -60,10 +60,10 @@ ft_findpoint(double pt, double *lims, int maxp, int minp, bool islog)
     if (islog) {
         tl = mylog10(lims[0]);
         th = mylog10(lims[1]);
-        return (((mylog10(pt) - tl) / (th - tl)) *
+        return (int)(((mylog10(pt) - tl) / (th - tl)) *
                 (maxp - minp) + minp);
     } else {
-        return (((pt - lims[0]) / (lims[1] - lims[0])) *
+        return (int)(((pt - lims[0]) / (lims[1] - lims[0])) *
                 (maxp - minp) + minp);
     }
 }
